pyproject.toml
[project]
name = "drgxcoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "flagembedding>=1.3.5",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langfuse>=3.10.1",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pinecone>=8.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
